requirements.dashboard.txt
requests>=2.31.0
certifi>=2026.01.04
jsonschema>=4.22.0
pyyaml>=6.0.1
python-dateutil>=2.9.0.post0
psycopg[binary]
lxml>=5.2.2
python-json-logger>=2.0.0

fastapi>=0.115.0
uvicorn[standard]>=0.32.0

openai>=1.0.0
anthropic>=0.39.0

httpx>=0.27.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0